backend/requirements.txt
addict==2.4.0
catboost==1.2.3
certifi==2025.1.31
charset-normalizer==3.3.2
contourpy==1.3.1
cycler==0.11.0
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.12.0
idna==3.7
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mmcv==2.0.0rc4
mmdet==3.3.0
mmengine==0.10.4
mmyolo==0.6.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==3.10.0
plotly==5.24.1
pooch==1.8.2
prettytable==3.16.0
pycocotools==2.0.8
pygments==2.19.1
pyparsing==3.2.0
pysocks==1.7.1
python-dateutil==2.9.0
pytz==2024.1
pyyaml==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.3.0
scipy==1.10.1
setuptools==75.8.0
shapely==2.0.7
six==1.16.0
sympy==1.13.3
tenacity==9.0.0
termcolor==2.5.0
terminaltables==3.1.10
threadpoolctl==3.6.0
torch==2.0.0
tqdm==4.67.1
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.43.0
yapf==0.43.0